+++ CHANGELOG.rst
@@ -4,13 +4,14 @@ Toolium Changelog
 v1.1.3
 ------
 
-*In development*
+*Release date: 2016-11-18*
 
 - Video download works in Selenium Grid 3
 - New config property 'binary' in [Firefox] section to configure the firefox binary path
 - Allow to configure visual baseline directory in ConfigFiles class (default: output/visualtests/baseline)
 - Delete IE and Edge cookies after tests
 - Fix wait_until_element_visible and wait_until_element_not_visible methods when the page element has a parent element
+- Add *imagemagick* as visual engine to have better diff images
 
 v1.1.2
 ------